requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain>=0.2.15
langchain-core>=0.2.37
langchain-community>=0.2.12
langchain-google-genai>=1.0.7
